restructure the confidential transfer instructions
