+++ programs/token-2022/src/instructions/extensions/confidential_transfer/mod.rs
@@ -1,3 +1,23 @@
-pub mod instructions;
+pub mod apply_pending_balance;
+pub mod approve_account;
+pub mod configure_account;
+pub mod configure_account_with_registry;
+pub mod deposit;
+pub mod disable_confidential_credits;
+pub mod disable_non_confidential_credits;
+pub mod empty_account;
+pub mod enable_confidential_credits;
+pub mod enable_non_confidential_credits;
+pub mod initialize_mint;
+pub mod transfer;
+pub mod transfer_with_fee;
+pub mod update_mint;
+pub mod withdraw;
 
-pub use instructions::*;
+pub use {
+    apply_pending_balance::*, approve_account::*, configure_account::*,
+    configure_account_with_registry::*, deposit::*, disable_confidential_credits::*,
+    disable_non_confidential_credits::*, empty_account::*, enable_confidential_credits::*,
+    enable_non_confidential_credits::*, initialize_mint::*, transfer::*, transfer_with_fee::*,
+    update_mint::*, withdraw::*,
+};
